+++ react/components/Spinner/index.js
@@ -2,6 +2,9 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import config from 'vtex-tachyons/config.json'
 
+const radius = 40
+const circ = 2 * radius * Math.PI
+
 const Spinner = ({ secondary, size }) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
@@ -14,13 +17,22 @@ const Spinner = ({ secondary, size }) => (
       cx="50"
       cy="50"
       fill="none"
+      r={radius}
       stroke={secondary ? config.colors.white : config.colors.blue}
       strokeWidth="10"
+      strokeDasharray={`0 0 2 ${circ}`}
       strokeLinecap="round"
-      r="40"
-      strokeDasharray="188.49555921538757 64.83185307179586"
+      strokeDashoffset="1"
       transform="rotate(96 50 50)"
     >
+      <animate
+        attributeName="stroke-dasharray"
+        dur="1.5s"
+        repeatCount="indefinite"
+        values={`0 0 2 ${circ};
+          0 0 ${circ * 0.75} ${circ};
+          0 ${circ - 2} ${circ * 0.75} ${circ}`}
+      />
       <animateTransform
         attributeName="transform"
         type="rotate"